Simply Tasmota Tally light support
Simple support for using Tasmota devices with an RGB LED as a tally
light. This adds a callback for OBS Frontend event so that scene changes
can be detected, and tally light status be updated when it happens.

Any source can be listed as a tally source in the configuration dialog.
When a named source is activated or deactivated in either the preview or
program views, an MQTT update is sent to cmnd/<source name>/color with
one of:
- ff0000 (red) for active in program view
- 00ff00 (green) for active in preview view
- 000000 (off) for not visible

Signed-off-by: Grant Likely <grant.likely@arm.com>